+++ R/add_minimum_distance.R
@@ -3,7 +3,7 @@
 #' @description \code{add_minimum_distance()} allows you to add the minimum
 #' distance (in kilometers) to a dyad-year or state-year data frame. These estimates
 #' are recorded in the \code{cow_mindist} and \code{gw_mindist} data that come with this package. The
-#' data are current as of the end of 2015.
+#' data are current as of the end of 2019.
 #'
 #'
 #' @return \code{add_minimum_distance()} takes a dyad-year or state-year data frame and adds the
@@ -19,6 +19,9 @@
 #' @param data a dyad-year data frame (either "directed" or "non-directed") or state-year data frame
 #'
 #' @references
+#'
+#' Schvitz, Guy et al. 2021. "Mapping The International System, 1886-2017: The \code{CShapes} 2.0 Dataset."
+#'  \emph{Journal of Conflict Resolution}. \url{https://journals.sagepub.com/doi/full/10.1177/00220027211013563}
 #'
 #' Weidmann, Nils B. and Kristian Skrede Gleditsch. 2010. "Mapping and Measuring Country Shapes: The \code{cshapes} Package." \emph{The R Journal} 2(1): 18-24.
 #'